requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
requests==2.25.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
